+++ src/CategoryApp/categoryService.ts
@@ -1,36 +1,7 @@
 import { create } from "ts-node";
 import categoryRepository from "./categoryRepository";
 import { Prisma } from "@prisma/client";
-
-type Product = Prisma.ProductGetPayload<{}>
-
-type Category = Prisma.CategoryGetPayload<{}>
-
-type CategoryWithProducts = Prisma.CategoryGetPayload<{
-    include: {
-        Products: true
-    }
-}>
-// {id: number, name: string, Products: [{name: string, id: number}]}
-
-interface ICategoryError{
-    status: 'error',
-    message: string
-}
-
-interface ICategoriesSuccess{
-    status: 'success',
-    data: Category[]
-}
-
-interface ICategorySuccess{
-    status: 'success',
-    data: Category
-}
-interface ICategoryWithProductsSuccess{
-    status: 'success',
-    data: CategoryWithProducts
-}
+import { ICategoriesSuccess, ICategoryError, ICategorySuccess, ICategoryWithProductsSuccess } from "./types";
 
 async function getAllCategories(): Promise< ICategoryError | ICategoriesSuccess > {
     const categories = await categoryRepository.getAllCategories();
